requirements.txt
viam-sdk
adafruit_mcp3xxx
busio
digitalio
board